+++ packages/react-magma-dom/src/components/Breadcrumb/index.tsx
@@ -2,9 +2,7 @@ import * as React from 'react';
 import styled from '../../theme/styled';
 import { ThemeContext } from '../../theme/ThemeContext';
 
-export interface BreadcrumbProps
-  extends React.HTMLAttributes<HTMLOListElement> {
-  ariaLabel?: string;
+export interface BreadcrumbProps extends React.HTMLAttributes<HTMLElement> {
   inverse?: boolean;
   minWidthToShow?: number;
   testId?: string;
@@ -40,7 +38,14 @@ export const Breadcrumb: React.FunctionComponent<
   BreadcrumbProps
 > = React.forwardRef(
   (
-    { ariaLabel, children, inverse, minWidthToShow, testId }: BreadcrumbProps,
+    {
+      'aria-label': ariaLabel,
+      children,
+      inverse,
+      minWidthToShow,
+      testId,
+      ...other
+    }: BreadcrumbProps,
     ref: any
   ) => {
     const theme = React.useContext(ThemeContext);
@@ -48,6 +53,7 @@ export const Breadcrumb: React.FunctionComponent<
     return (
       <BreadCrumbContext.Provider value={{ inverse }}>
         <StyledNav
+          {...other}
           aria-label={ariaLabel ? ariaLabel : 'Breadcrumb'}
           breakpoint={minWidthToShow ? `${minWidthToShow}px` : '0'}
           data-testid={testId}